requirements.txt
pandas==0.24.2
opencv-contrib-python-headless==4.4.0.44
numpy==1.16.4
Pillow==8.0.0
streamlit==0.69.2
